fixing segfault on freeObjectTable
When the object table is empty, freeObjectTable tried to free a
table of size 1, instead of doing nothing. This is now fixed.
This case could happen with corrupted emf files.

+++ inc/EMFSVG_private.h
@@ -223,7 +223,11 @@ extern "C" {
         bool inPath;
         // object table
         emfGraphObject * objectTable;
-        uint16_t objectTableSize; 
+        // size of the object table (warning, could be negative)
+        // initialized to -1
+        // the "real" object table is objectTableSize + 1 
+        // reason: (indexes in emf files start a 1 and not 0)
+        uint32_t objectTableSize; 
         // scaling ratio
         double scaling;
         double scalingX;
